CASMCMS-8060 Fix incomplete non-root cray-cfs-api changes
When the "cray-cfs-api" chart was converted to non-root, the
change was incomplete.  In addition to specifying a
"securityContext" for the container, a "securityContext" also
needs to be specified for the pod.  This is what tells
Kubernetes how to set up permissions on PVCs.  Without that,
the PVCs will have the default permissions of the underlying
storage class, which can vary between storage classes.

The only reason things are currently working is because the
CephFS storage class defaults to a top level directory of
"root/root", mode "0x777".  In Mercury we switch to a cStor
storage class, where the default permissions are "0x755", and
non-root containers are not able to create files.