Update CHANGELOG.md
Even though this consists solely of a bug fix, I am bumping the minor number, because this is being targeted for CSM 1.6.1, and is not being pulled into CSM 1,.6.0. On the off chance that we find some CFS problem that we DO need to fix in CSM 1.6.0, I want to leave "room" for that version number-wise.